% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbonace.R
\name{carbonace}
\alias{carbonace}
\title{carbonace shinyapp}
\usage{
carbonace(init = "x <- 1", viewer = shiny::browserViewer())
}
\arguments{
\item{init}{character, intial script to initialize shinyAce with.}

\item{viewer}{which \link[shiny:viewer]{viewer} to open the app in.}
}
\description{
shinyapp to convert shinyAce to images for code sharing
}
\details{
The app has an RStudio addin, which you can highlight text in the editor
and the app will initialize with the contents.
}
\examples{
if(interactive()){
 carbonace:::carbonace()
 }
}
